+++ @ecomplus/storefront-app/src/components/js/EcSummary.js
@@ -1,4 +1,5 @@
 import {
+  // i19balanceOrPoints,
   i19buyer,
   i19contactPhone,
   i19discount,
@@ -52,6 +53,10 @@ export default {
   },
 
   computed: {
+    i19balanceOrPoints: () => i18n({
+      en_us: 'Balance or points',
+      pt_br: 'Saldo ou pontos'
+    }),
     i19buyer: () => i18n(i19buyer),
     i19contactPhone: () => i18n(i19contactPhone),
     i19discount: () => i18n(i19discount),