+++ @ecomplus/storefront-template/template/scss/themes/ecom-beauty/config.js
@@ -1,0 +1,7 @@
+const themeConfig = {
+  product: {
+    hasQuantitySelector: true,
+    hasPromotionTimer: false
+  }
+}
+module.exports = themeConfig