+++ @ecomplus/storefront-starter/CHANGELOG.md
@@ -3,6 +3,134 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-starter.23](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.22...@ecomplus/storefront-starter@2.0.0-starter.23) (2022-07-28)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.22](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.21...@ecomplus/storefront-starter@2.0.0-starter.22) (2022-07-26)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.21](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.20...@ecomplus/storefront-starter@2.0.0-starter.21) (2022-07-16)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.20](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.19...@ecomplus/storefront-starter@2.0.0-starter.20) (2022-07-06)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.19](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.18...@ecomplus/storefront-starter@2.0.0-starter.19) (2022-07-05)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.18](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.17...@ecomplus/storefront-starter@2.0.0-starter.18) (2022-07-04)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.17](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.16...@ecomplus/storefront-starter@2.0.0-starter.17) (2022-06-03)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.16](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.15...@ecomplus/storefront-starter@2.0.0-starter.16) (2022-05-25)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.15](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.14...@ecomplus/storefront-starter@2.0.0-starter.15) (2022-05-18)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.14](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.13...@ecomplus/storefront-starter@2.0.0-starter.14) (2022-05-17)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.13](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.12...@ecomplus/storefront-starter@2.0.0-starter.13) (2022-05-17)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.12](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.11...@ecomplus/storefront-starter@2.0.0-starter.12) (2022-05-15)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.11](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.10...@ecomplus/storefront-starter@2.0.0-starter.11) (2022-05-14)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.10](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.9...@ecomplus/storefront-starter@2.0.0-starter.10) (2022-05-11)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.9](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.8...@ecomplus/storefront-starter@2.0.0-starter.9) (2022-05-06)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
+# [2.0.0-starter.8](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.7...@ecomplus/storefront-starter@2.0.0-starter.8) (2022-04-21)
+
+**Note:** Version bump only for package @ecomplus/storefront-starter
+
+
+
+
+
 # [2.0.0-starter.7](https://github.com/ecomplus/storefront-starter/compare/@ecomplus/storefront-starter@2.0.0-starter.6...@ecomplus/storefront-starter@2.0.0-starter.7) (2022-04-01)
 
 **Note:** Version bump only for package @ecomplus/storefront-starter